+++ elasticsearch5/src/test/java/org/vertexium/elasticsearch5/Elasticsearch5SearchIndexTest.java
@@ -1,14 +1,70 @@
 package org.vertexium.elasticsearch5;
 
+import org.junit.Assert;
+import org.junit.Before;
 import org.junit.ClassRule;
+import org.junit.Test;
+import org.vertexium.*;
+import org.vertexium.inmemory.InMemoryAuthorizations;
+import org.vertexium.inmemory.InMemoryGraph;
+import org.vertexium.inmemory.InMemoryGraphConfiguration;
+import org.vertexium.query.QueryResultsIterable;
+import org.vertexium.query.SortDirection;
+import org.vertexium.test.GraphTestBase;
 
-public class Elasticsearch5SearchIndexTest extends Elasticsearch5SearchIndexTestBase {
+import static org.vertexium.util.IterableUtils.count;
+
+public class Elasticsearch5SearchIndexTest extends GraphTestBase {
 
     @ClassRule
     public static ElasticsearchResource elasticsearchResource = new ElasticsearchResource();
 
     @Override
-    protected ElasticsearchResource getElasticsearchResource() {
-        return elasticsearchResource;
+    protected Authorizations createAuthorizations(String... auths) {
+        return new InMemoryAuthorizations(auths);
+    }
+
+    @Before
+    @Override
+    public void before() throws Exception {
+        elasticsearchResource.dropIndices();
+        super.before();
+    }
+
+    @Override
+    protected Graph createGraph() throws Exception {
+        InMemoryGraphConfiguration configuration = new InMemoryGraphConfiguration(elasticsearchResource.createConfig());
+        return InMemoryGraph.create(configuration);
+    }
+
+    private Elasticsearch5SearchIndex getSearchIndex() {
+        return (Elasticsearch5SearchIndex) ((GraphWithSearchIndex) graph).getSearchIndex();
+    }
+
+    protected boolean isFieldNamesInQuerySupported() {
+        return true;
+    }
+
+    @Override
+    protected boolean disableEdgeIndexing(Graph graph) {
+        Elasticsearch5SearchIndex searchIndex = (Elasticsearch5SearchIndex) ((GraphWithSearchIndex) graph).getSearchIndex();
+        searchIndex.getConfig().getGraphConfiguration().set(GraphConfiguration.SEARCH_INDEX_PROP_PREFIX + "." + ElasticsearchSearchIndexConfiguration.INDEX_EDGES, "false");
+        return true;
+    }
+
+    @Override
+    protected boolean isLuceneQueriesSupported() {
+        return true;
+    }
+
+    @Test
+    @Override
+    public void testGraphQuerySortOnPropertyThatHasNoValuesInTheIndex() {
+        super.testGraphQuerySortOnPropertyThatHasNoValuesInTheIndex();
+
+        getSearchIndex().clearIndexInfoCache();
+
+        QueryResultsIterable<Vertex> vertices = graph.query(AUTHORIZATIONS_A).sort("age", SortDirection.ASCENDING).vertices();
+        Assert.assertEquals(2, count(vertices));
     }
 }