+++ sql/src/test/java/org/vertexium/sql/SqlGraphTest.java
@@ -63,4 +63,9 @@ public class SqlGraphTest extends GraphTestBase {
     protected Authorizations createAuthorizations(String... auths) {
         return new InMemoryAuthorizations(auths);
     }
+
+    @Override
+    protected boolean isAdvancedGeoQuerySupported() {
+        return false;
+    }
 }